added dlsegkidneyfatwater, with test, model.rst card and nnunet .png

+++ src/miblab/data.py
@@ -142,11 +142,10 @@ def zenodo_fetch(dataset: str, folder: str, doi: str = None, filename: str = Non
     return extract_to
 
     
-
-
 def clear_cache_datafiles(directory: str, verbose: bool = True):
     """
-    Delete all files and subdirectories in the specified cache directory.
+    Delete all files and subdirectories in the specified cache directory,
+    except for '__init__' files.
 
     Args:
         directory (str): Path to the directory to clear.
@@ -162,6 +161,11 @@ def clear_cache_datafiles(directory: str, verbose: bool = True):
     deleted = []
     for item in os.listdir(directory):
         path = os.path.join(directory, item)
+
+        # Skip __init__ files (e.g., __init__.py, __init__.pyc)
+        if os.path.isfile(path) and os.path.splitext(item)[0] == '__init__':
+            continue
+
         try:
             if os.path.isfile(path) or os.path.islink(path):
                 os.remove(path)
@@ -175,7 +179,7 @@ def clear_cache_datafiles(directory: str, verbose: bool = True):
                     print(f"Deleted folder: {path}")
         except Exception as e:
             print(f"Error deleting {path}: {e}")
-    
+
     if verbose and not deleted:
         print("Directory is already clean.")
